chore: run cargo clippy --fix

+++ crates/aoc_2019/src/aoc/day01.rs
@@ -22,7 +22,7 @@ impl AoCProblem for Day01 {
     fn part2(&mut self) -> Solution {
         self.modules.iter().fold(0, |acc, &x| {
             let mut mass = x as i64;
-            let mut res = acc as i64;
+            let mut res = acc;
             loop {
                 mass = mass / 3 - 2;
                 if mass <= 0 {
